+++ .env
@@ -1,0 +1,1 @@
+CONNECTION_STRING=mongodb+srv://dshpak:Qweszxc11@cluster.1ppmqw0.mongodb.net/db-contacts
