+++ .vscode/settings.json
@@ -5,20 +5,7 @@
         "**/.hg": true,
         "**/CVS": true,
         "**/.DS_Store": true,
-        "**/Thumbs.db": true,
-        "contacts.mongodb": true,
-        "package-lock.json": true,
-        ".eslintignore": true,
-        ".eslintrc.js": true,
-        "node_modules": true,
-        ".vscode": true
+        "**/Thumbs.db": true
     },
-    "hide-files.files": [
-        "contacts.mongodb",
-        "package-lock.json",
-        ".eslintignore",
-        ".eslintrc.js",
-        "node_modules",
-        ".vscode"
-    ]
+    "hide-files.files": []
 }
