+++ contacts.mongodb
@@ -1,65 +1,65 @@
-use('test')
+use('db-contacts')
 
-db.contacts.findOneAndDelete({name: "Allen Raymond"})
-// db.contacts.insertMany([
-//   {
-//     "name": "Allen Raymond",
-//     "email": "nulla.ante@vestibul.co.uk",
-//     "phone": "(992) 914-3792",
-//     "favorite": false
-//   },
-//   {
-//     "name": "Chaim Lewis",
-//     "email": "dui.in@egetlacus.ca",
-//     "phone": "(294) 840-6685",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Kennedy Lane",
-//     "email": "mattis.Cras@nonenimMauris.net",
-//     "phone": "(542) 451-7038",
-//     "favorite": false
-//   },
-//   {
-//     "name": "Wylie Pope",
-//     "email": "est@utquamvel.net",
-//     "phone": "(692) 802-2949",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Cyrus Jackson",
-//     "email": "nibh@semsempererat.com",
-//     "phone": "(501) 472-5218",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Abbot Franks",
-//     "email": "scelerisque@magnis.org",
-//     "phone": "(186) 568-3720",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Reuben Henry",
-//     "email": "pharetra.ut@dictum.co.uk",
-//     "phone": "(715) 598-5792",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Simon Morton",
-//     "email": "dui.Fusce.diam@Donec.com",
-//     "phone": "(233) 738-2360",
-//     "favorite": true
-//   },
-//   {
-//     "name": "Thomas Lucas",
-//     "email": "nec@Nulla.com",
-//     "phone": "(704) 398-7993",
-//     "favorite": false
-//   },
-//   {
-//     "name": "Alec Howard",
-//     "email": "Donec.elementum@scelerisquescelerisquedui.net",
-//     "phone": "(748) 206-2688",
-//     "favorite": true
-//   }
-// ])
+// db.contacts.findOneAndDelete({name: "Allen Raymond"})
+db.contacts.insertMany([
+  {
+    "name": "Allen Raymond",
+    "email": "nulla.ante@vestibul.co.uk",
+    "phone": "(992) 914-3792",
+    "favorite": false
+  },
+  {
+    "name": "Chaim Lewis",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685",
+    "favorite": true
+  },
+  {
+    "name": "Kennedy Lane",
+    "email": "mattis.Cras@nonenimMauris.net",
+    "phone": "(542) 451-7038",
+    "favorite": false
+  },
+  {
+    "name": "Wylie Pope",
+    "email": "est@utquamvel.net",
+    "phone": "(692) 802-2949",
+    "favorite": true
+  },
+  {
+    "name": "Cyrus Jackson",
+    "email": "nibh@semsempererat.com",
+    "phone": "(501) 472-5218",
+    "favorite": true
+  },
+  {
+    "name": "Abbot Franks",
+    "email": "scelerisque@magnis.org",
+    "phone": "(186) 568-3720",
+    "favorite": true
+  },
+  {
+    "name": "Reuben Henry",
+    "email": "pharetra.ut@dictum.co.uk",
+    "phone": "(715) 598-5792",
+    "favorite": true
+  },
+  {
+    "name": "Simon Morton",
+    "email": "dui.Fusce.diam@Donec.com",
+    "phone": "(233) 738-2360",
+    "favorite": true
+  },
+  {
+    "name": "Thomas Lucas",
+    "email": "nec@Nulla.com",
+    "phone": "(704) 398-7993",
+    "favorite": false
+  },
+  {
+    "name": "Alec Howard",
+    "email": "Donec.elementum@scelerisquescelerisquedui.net",
+    "phone": "(748) 206-2688",
+    "favorite": true
+  }
+])
