+++ models/contacts/db-access-service.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const Contact = require("./contacts-mongodb-schema");
+const Contact = require("./db-contacts-schema");
 
 async function listContacts() {
     return Contact.find();
@@ -10,8 +10,8 @@ async function getContactById(contactId) {
     return Contact.findById(contactId);
 }
 
-async function addContact({ name, email, phone, favorite }) {
-    return Contact.create({ name, email, phone, favorite });
+async function addContact(newContactObj) {
+    return Contact.create(newContactObj);
 }
 
 async function updateContact(contactId, newDataObj) {
@@ -27,13 +27,13 @@ async function removeContact(contactId) {
     return Contact.findByIdAndRemove(contactId);
 }
 
-// async function updateStatusContact(contactId, newDataObj) {
-//     if (!mongoose.Types.ObjectId.isValid(contactId)) return null;
-//     return Contact.findByIdAndUpdate(contactId, newDataObj, {
-//         new: true,
-//         runValidators: true,
-//     });
-// }
+async function updateStatusContact(contactId, newDataObj) {
+    if (!mongoose.Types.ObjectId.isValid(contactId)) return null;
+    return Contact.findByIdAndUpdate(contactId, newDataObj, {
+        new: true,
+        runValidators: true,
+    });
+}
 
 module.exports = {
     listContacts,
@@ -41,5 +41,5 @@ module.exports = {
     addContact,
     updateContact,
     removeContact,
-    // updateStatusContact,
+    updateStatusContact,
 };
